+++ src/components/Menu/index.tsx
@@ -1,0 +1,10 @@
+export interface MenuProps {
+}
+
+export function Menu({ }: MenuProps) {
+  return (
+    <div>
+      <p>Menu</p>
+    </div>
+  )
+}