+++ src/components/MenuItem/index.tsx
@@ -1,0 +1,10 @@
+export interface MenuItemProps {
+}
+
+export function MenuItem({ }: MenuItemProps) {
+  return (
+    <div>
+      <p>Menu Item</p>
+    </div>
+  )
+}