+++ src/containers/Reddit/reducer.ts
@@ -2,19 +2,21 @@
 import { set as _set } from 'dot-prop-immutable';
 
 // Types
-import { IRedditTop50Root } from '../../constants/services/reddit/types';
+import { IRedditTop50DataItem } from '../../constants/services/reddit/types';
 import { RedditActionTypes } from './actionsTypes';
 import { IRedditFetchActions } from './actions';
 
 export interface IRedditState {
     data: {
-        top50: IRedditTop50Root | undefined
+        top50: IRedditTop50DataItem[]
+        visitedList: string[]
     }
 }
 
 const initialState: IRedditState = {
     data: {
-        top50: undefined
+        top50: [],
+        visitedList: [],
     }
 };
 
@@ -22,7 +24,7 @@ export function reducer(state: IRedditState = initialState, action: IRedditFetch
     const { type, payload } = action;
 
     if (type === RedditActionTypes.FETCH_DATA) {
-        state = _set(state, 'data.top50', payload);
+        state = _set(state, 'data.top50', payload.data.children);
     }
 
     return state;