+++ src/containers/Reddit/selectors.ts
@@ -1,0 +1,8 @@
+import { useSelector } from 'react-redux';
+import { RootState } from '../../providers/redux/rootReducer';
+
+export function usePostList() {
+    const { data } = useSelector((store: RootState) => store.REDDIT);
+
+    return data.top50;
+}